SPEED SUPPORT AND CHANGES TO THE PROCESS METHODS
Speed support has been added, being possible to both get it and set it.
In order to achieve this, some changes to the engine methods (_sendOrder, _buildPath and _xmlParse) has been made. Now they are more generic, so they can receive an object with parameters for the request and also the responsibility for process the information has been moved up.